# frontend/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "paths" : {
...
    "/api/v1/admin/impact-rules/{kind}" : {
      "put" : {
        "tags" : [ "impact-rules-admin-controller" ],
        "operationId" : "updateRule",
        "parameters" : [ {
          "name" : "kind",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "CALL", "INSTANTIATION", "METHOD_REF", "TYPE_REF", "EXTENDS", "IMPLEMENTS", "OVERRIDES", "FIELD_READ", "FIELD_WRITE", "ANNOTATION" ]
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RuleChange"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImpactRule"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/entry-point-annotations/{id}" : {
      "put" : {
        "tags" : [ "impact-rules-admin-controller" ],
        "operationId" : "update_3",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AnnotationChange"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntryPointAnnotation"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "impact-rules-admin-controller" ],
        "operationId" : "delete_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/admin/artifact-repositories/{id}" : {
      "get" : {
        "tags" : [ "artifact-repository-admin-controller" ],
        "operationId" : "get_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArtifactRepositoryView"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "artifact-repository-admin-controller" ],
        "operationId" : "update_4",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ArtifactRepositoryUpdate"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArtifactRepositoryView"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "artifact-repository-admin-controller" ],
        "operationId" : "delete_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/index/runs" : {
      "get" : {
        "tags" : [ "index-run-controller" ],
        "operationId" : "list",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageIndexRunSummary"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "index-run-controller" ],
        "operationId" : "start",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StartRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StartedRun"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/index/runs/{runId}/cancel" : {
      "post" : {
        "tags" : [ "index-run-controller" ],
        "operationId" : "cancel",
        "parameters" : [ {
          "name" : "runId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/v1/admin/entry-point-annotations" : {
      "get" : {
        "tags" : [ "impact-rules-admin-controller" ],
        "operationId" : "annotations",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/EntryPointAnnotation"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "impact-rules-admin-controller" ],
        "operationId" : "create_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AnnotationChange"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntryPointAnnotation"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/artifact-repositories" : {
      "get" : {
        "tags" : [ "artifact-repository-admin-controller" ],
        "operationId" : "list_3",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ArtifactRepositoryView"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "artifact-repository-admin-controller" ],
        "operationId" : "create_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ArtifactRepositoryUpdate"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArtifactRepositoryView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/artifact-repositories/{id}/test" : {
      "post" : {
        "tags" : [ "artifact-repository-admin-controller" ],
        "operationId" : "test_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "boolean"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/repositories/{id}/runs" : {
      "get" : {
        "tags" : [ "index-run-controller" ],
        "operationId" : "repositoryRuns",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageIndexRunRepoView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/index/runs/{runId}" : {
      "get" : {
        "tags" : [ "index-run-controller" ],
        "operationId" : "get_5",
        "parameters" : [ {
          "name" : "runId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IndexRunView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "RuleChange" : {
        "type" : "object",
        "properties" : {
          "propagates" : {
            "type" : "boolean"
          },
          "shownAtLevel1" : {
            "type" : "boolean"
          }
        }
      },
      "ImpactRule" : {
        "type" : "object",
        "properties" : {
          "kind" : {
            "type" : "string",
            "enum" : [ "CALL", "INSTANTIATION", "METHOD_REF", "TYPE_REF", "EXTENDS", "IMPLEMENTS", "OVERRIDES", "FIELD_READ", "FIELD_WRITE", "ANNOTATION" ]
          },
          "propagates" : {
            "type" : "boolean"
          },
          "shownAtLevel1" : {
            "type" : "boolean"
          }
        }
      },
      "AnnotationChange" : {
        "type" : "object",
        "properties" : {
          "annotationFqn" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          }
        }
      },
      "EntryPointAnnotation" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "annotationFqn" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          }
        }
      },
      "ArtifactRepositoryUpdate" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "secret" : {
            "type" : "string"
          },
          "mirrorOf" : {
            "type" : "string"
          },
          "sortOrder" : {
            "type" : "integer",
            "format" : "int32"
          },
          "enabled" : {
            "type" : "boolean"
          }
        }
      },
      "ArtifactRepositoryView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "secretSet" : {
            "type" : "boolean"
          },
          "mirrorOf" : {
            "type" : "string"
          },
          "sortOrder" : {
            "type" : "integer",
            "format" : "int32"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "lastTestStatus" : {
            "type" : "string"
          },
          "lastTestAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "StartRequest" : {
        "type" : "object",
        "properties" : {
          "scope" : {
            "type" : "string",
            "enum" : [ "ALL", "CONNECTION", "REPOSITORY" ]
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "force" : {
            "type" : "boolean"
          }
        }
      },
      "StartedRun" : {
        "type" : "object",
        "properties" : {
          "runId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "RepositoryRef" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "projectKey" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          }
        }
      },
...
      "IndexRunRepoView" : {
        "type" : "object",
        "properties" : {
          "runId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "repository" : {
            "$ref" : "#/components/schemas/RepositoryRef"
          },
          "commit" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "classpathMode" : {
            "type" : "string"
          },
          "error" : {
            "type" : "string"
          },
          "symbolCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "usageCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "warningCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "durationMs" : {
            "type" : "integer",
            "format" : "int64"
          },
          "finishedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "artifactInstall" : {
            "type" : "string"
          },
          "artifactInstallError" : {
            "type" : "string"
          }
        }
      },
      "PageIndexRunRepoView" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IndexRunRepoView"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "IndexRunSummary" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "trigger" : {
            "type" : "string"
          },
          "scope" : {
            "type" : "string"
          },
          "scopeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "status" : {
            "type" : "string"
          },
          "startedBy" : {
            "type" : "string"
          },
          "startedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "finishedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "cancelRequested" : {
            "type" : "boolean"
          }
        }
      },
      "PageIndexRunSummary" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IndexRunSummary"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "IndexRunView" : {
        "type" : "object",
        "properties" : {
          "run" : {
            "$ref" : "#/components/schemas/IndexRunSummary"
          },
          "error" : {
            "type" : "string"
          },
          "repositoriesByStatus" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int32"
            }
          },
          "repositories" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IndexRunRepoView"
            }
          },
          "inProgress" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RepositoryRef"
            }
          }
        }
      },
...
    }
  }
}